requirements.txt
Django==2.2.7
django-service-objects==0.6.0
djangorestframework==3.11.0
pkg-resources==0.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
